+++ src/App.jsx
@@ -5,19 +5,19 @@ import { Courses } from './components/Courses/Courses';
 import { CreateCourse } from './components/CreateCourse/CreateCourse';
 import { DefaultPage } from './components/DefaultPage/DefaultPage';
 
-import { useEffect, useState } from 'react';
-
-import { mockedAuthorsList } from './constants';
-
-import { getCurrentDate } from './helpers/dateGeneratop';
+import { useEffect } from 'react';
 
 import { Route, Routes } from 'react-router-dom';
 import { Registration } from './components/Registration/Registration';
 import { Login } from './components/Login/Login';
 import { CourseInfo } from './components/CourseInfo/CourseInfo';
 import axios from 'axios';
-import { useDispatch } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 import { setUser } from './store/user/actionCreators';
+import { api } from './servisces';
+import { setCourses } from './store/courses/actionCreators';
+import { setAuthors } from './store/authors/actionCreators';
+import { selectUser } from './store';
 
 axios.defaults.baseURL = 'http://localhost:4000';
 axios.interceptors.request.use((config) => {
@@ -28,9 +28,8 @@ axios.interceptors.request.use((config) => {
 });
 
 const App = () => {
-	const [authorsList, setAuthors] = useState(mockedAuthorsList);
-	const [isNewCourseOpen, setNewCourseOpen] = useState(false);
 	const dispatch = useDispatch();
+	const currentUser = useSelector(selectUser);
 
 	useEffect(() => {
 		const user = localStorage.getItem('userData');
@@ -39,52 +38,23 @@ const App = () => {
 		}
 	}, [dispatch]);
 
-	const createNewAuthor = (name) => {
-		const author = {
-			id: uuidv4(),
-			name: name,
-		};
-
-		setAuthors([...authorsList, author]);
-	};
-
-	const createNewCourse = (title, description, duration, authors) => {
-		const course = {
-			id: uuidv4(),
-			title: title,
-			description: description,
-			creationDate: getCurrentDate(),
-			duration: duration,
-			authors: authors,
-		};
-
-		// setCourses([...coursesList, course]);
-		setNewCourseOpen(false);
-	};
+	useEffect(() => {
+		if (currentUser.isAuth) {
+			api.courses.getAllCourses().then((response) => {
+				dispatch(setCourses(response.data.result));
+			});
+			api.authors.getAllAuthors().then((response) => {
+				dispatch(setAuthors(response.data.result));
+			});
+		}
+	}, [currentUser, dispatch]);
 
 	return (
 		<div>
 			<Header />
 			<Routes>
-				<Route
-					path='/courses'
-					element={
-						<Courses
-							onAddNewCourse={() => setNewCourseOpen(true)}
-							authorsList={authorsList}
-						/>
-					}
-				></Route>
-				<Route
-					path='/courses/add'
-					element={
-						<CreateCourse
-							createNewAuthor={createNewAuthor}
-							createNewCourse={createNewCourse}
-							authorsList={authorsList}
-						/>
-					}
-				></Route>
+				<Route path='/courses' element={<Courses />}></Route>
+				<Route path='/courses/add' element={<CreateCourse />}></Route>
 				<Route path='/registration' element={<Registration />}></Route>
 				<Route path='/login' element={<Login />}></Route>
 				<Route path='/courses/:courseId' element={<CourseInfo />}></Route>
